+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.64] - 2022-09-09
+### Fixed
+- Collect Earth: calculated attribute sending EXTRA_attrName values;
+
+### Added
+- Data export: include sampling design item label (if defined);
+
 ## [4.0.63] - 2022-06-30
 ### Fixed
 - Removed dependency on deprecated Apache Commons DDLUtils;